Make sure incrementing and decrementing reference counts happens atomically.
* \sscli\tests\refcounting\refcounting.vcproj
* \sscli\tests\refcounting\rsources
* \sscli\tests\refcounting\TestConcurrency.cs
	New test.

* \sscli\clr\src\vm\refcount.cpp
* \sscli\clr\src\vm\refcount.h
	Make m_RefCount a LONG. 
	(Release) - Use InterlockedDecrement.
	(AddRef) - Use InterlockedIncrement.


git-svn-id: svn://localhost@153 23913781-f3ab-ca43-b79f-3fa6c8d344c4

+++ trunk/sscli/clr/src/vm/refcount.h
@@ -31,7 +31,7 @@ public:
     Object* GetObject();
 private:
     void Finalize();
-    int m_RefCount;
+    LONG m_RefCount;
 };
 
 OBJECTREF RCAllocateObject( MethodTable* pMT );
